pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tripgenie"
version = "0.1.0"
description = "TripGenie: Agentic Travel Planning & Booking Assistant"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "TripGenie Team" }]
keywords = ["travel", "ai", "llm", "agents", "fastapi", "genai"]

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "httpx>=0.27.0",
    "openai>=1.14.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "rich>=13.7.0",
    "streamlit>=1.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pre-commit>=3.7.0",
    "pandas-stubs>=2.2.0",
    "types-PyYAML>=6.0.12",
]

[project.scripts]
tripgenie = "app.main:main"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ---------------------------------------------------------------------------
# Ruff – linting + formatting
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "ANN"]
ignore = [
    "ANN101", "ANN102", "ANN401",
    "ANN001", "ANN002", "ANN003",   # relax annotation rules for tests
    "E501",                          # line length handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "S101"]
"notebooks/*" = ["ANN", "E402", "F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
warn_return_any = false
warn_unused_configs = true

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = ["ignore::DeprecationWarning"]

# ---------------------------------------------------------------------------
# coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source = ["app"]
omit = ["app/ui/*", "tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 60
